pyproject.toml
[tool.poetry]
name = "tellmewithouttellingme_excercise"
version = "0.1.0"
description = ""
authors = ["John Albert Balansag <johnalbert.balansag@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.13.0"
factory-boy = "^3.2.1"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
dependency-injector = "^4.39.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
